requirements.txt
Jinja2>=3.1
PyYAML>=6.0
livereload>=2.6
